bundle scripts: when copy apps for a bundle use separpate copies for fds and smokeview

+++ Firebot/copy_apps.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+type=$1
 
 #---------------------------------------------
 #                   MKDIR
@@ -64,24 +65,28 @@ TODIR=$HOME/.bundle
 
 # copy smokeview files
 
-MKDIR $TODIR/smv
-rm -f $TODIR/smv/*
-echo
-echo ***copying smokeview  apps
-CP $smvrepo/Build/background/intel$OS background   $TODIR/smv background
-CP $smvrepo/Build/dem2fds/intel$OS    dem2fds$OS   $TODIR/smv dem2fds
-CP $smvrepo/Build/hashfile/intel$OS   hashfile$OS  $TODIR/smv hashfile
-CP $smvrepo/Build/smokediff/intel$OS  smokediff$OS $TODIR/smv smokediff
-CP $smvrepo/Build/smokeview/intel$OS  smokeview$OS $TODIR/smv smokeview
-CP $smvrepo/Build/smokezip/intel$OS   smokezip$OS  $TODIR/smv smokezip
-CP $smvrepo/Build/wind2fds/intel$OS   wind2fds$OS  $TODIR/smv wind2fds
+if [ "$type" == "smv" ]; then
+  MKDIR $TODIR/smv
+  rm -f $TODIR/smv/*
+  echo
+  echo ***copying smokeview  apps
+  CP $smvrepo/Build/background/intel$OS background   $TODIR/smv background
+  CP $smvrepo/Build/dem2fds/intel$OS    dem2fds$OS   $TODIR/smv dem2fds
+  CP $smvrepo/Build/hashfile/intel$OS   hashfile$OS  $TODIR/smv hashfile
+  CP $smvrepo/Build/smokediff/intel$OS  smokediff$OS $TODIR/smv smokediff
+  CP $smvrepo/Build/smokeview/intel$OS  smokeview$OS $TODIR/smv smokeview
+  CP $smvrepo/Build/smokezip/intel$OS   smokezip$OS  $TODIR/smv smokezip
+  CP $smvrepo/Build/wind2fds/intel$OS   wind2fds$OS  $TODIR/smv wind2fds
+fi
 
 # copy fds files
 
-echo
-echo ***copying fds apps
-MKDIR $TODIR/fds
-rm -f $TODIR/fds/*
-CP $fdsrepo/Build/${MPI}_intel$OS               fds_${MPI}_intel$OS $TODIR/fds fds
-CP $fdsrepo/Utilities/fds2ascii/intel$OS        fds2ascii$OS        $TODIR/fds fds2ascii
-CP $fdsrepo/Utilities/test_mpi/${MPI}_intel$OS2 test_mpi            $TODIR/fds test_mpi
+if [ "$type" == "fds" ]; then
+  echo
+  echo ***copying fds apps
+  MKDIR $TODIR/fds
+  rm -f $TODIR/fds/*
+  CP $fdsrepo/Build/${MPI}_intel$OS               fds_${MPI}_intel$OS $TODIR/fds fds
+  CP $fdsrepo/Utilities/fds2ascii/intel$OS        fds2ascii$OS        $TODIR/fds fds2ascii
+  CP $fdsrepo/Utilities/test_mpi/${MPI}_intel$OS2 test_mpi            $TODIR/fds test_mpi
+fi
